+++ base/numerics/safe_math_arm_impl.h
@@ -56,7 +56,7 @@ struct ClampedAddFastAsmOp {
     // This will get promoted to an int, so let the compiler do whatever is
     // clever and rely on the saturated cast to bounds check.
     if (IsIntegerArithmeticSafe<int, T, U>::value)
-      return saturated_cast<V>(x + y);
+      return saturated_cast<V>(static_cast<int>(x) + static_cast<int>(y));
 
     int32_t result;
     int32_t x_i32 = checked_cast<int32_t>(x);
@@ -82,7 +82,7 @@ struct ClampedSubFastAsmOp {
     // This will get promoted to an int, so let the compiler do whatever is
     // clever and rely on the saturated cast to bounds check.
     if (IsIntegerArithmeticSafe<int, T, U>::value)
-      return saturated_cast<V>(x - y);
+      return saturated_cast<V>(static_cast<int>(x) - static_cast<int>(y));
 
     int32_t result;
     int32_t x_i32 = checked_cast<int32_t>(x);
