+++ url/url_canon_stdstring.h
@@ -46,7 +46,7 @@ class COMPONENT_EXPORT(URL) StdStringCanonOutput : public CanonOutput {
   // Must be called after writing has completed but before the string is used.
   void Complete();
 
-  void Resize(int sz) override;
+  void Resize(size_t sz) override;
 
  protected:
   // `str_` is not a raw_ptr<...> for performance reasons (based on analysis of
